% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBlock.R
\name{convertBlock}
\alias{convertBlock}
\alias{make_block_bamboo}
\alias{make_block_wheat}
\title{Functions to Convert Blocks to Specific Farm Types}
\usage{
make_block_bamboo(block)

make_block_wheat(block, grow_prob = 0.25)
}
\arguments{
\item{block}{\code{minecraft_block} object.}

\item{grow_prob}{Probability of growth when assigned a random tick.}
}
\description{
The functions provided here operate on a \code{minecraft_block}
object to create blocks of farmland.
}
