% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedCropOutput.R
\name{ExpectedCropOutput}
\alias{ExpectedCropOutput}
\alias{expected_crop_output}
\title{Expected Crop Output}
\usage{
expected_crop_output(
  yield,
  growth_probability,
  stages = 1,
  p = 0.5,
  method = c("median", "mean", "quantile"),
  unit = "minute"
)
}
\arguments{
\item{yield}{\code{numeric(1)}. The number of times returned by harvesting
a crop. This may be an average and need not be a whole number.}

\item{growth_probability}{\code{numeric(1)} on the interval (0, 1). The
probability that a crop will advance a growth stage when assigned a
random tick.}

\item{stages}{\code{integerish(1)}. The number of growth stages a crop requires
before becoming mature.}

\item{p}{\code{numeric(1)} on the interval (0, 1). The probability used for
determining the quartile of the probability distribution. Used only
when \code{unit = "minute"}.}

\item{method}{\code{character(1)}. One of \code{c("median", "mean", "quantile")}.
The method of calculation to be based on either the median, mean, or
quantile of the probability distribution.}

\item{unit}{\code{character(1)}. One of \code{c("tick", "second", "minute", "hour")}.
The unit of time for the metric.}
}
\description{
The Expected Crop Output (ECO) is a measure of how much crop
is produced by a single block of farm per minute.
}
