% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_quantile.R
\name{geom_quantile}
\alias{geom_quantile}
\title{Quantiles for Geometric Distribution}
\usage{
geom_quantile(p, growth_probability)
}
\arguments{
\item{p}{\code{numeric(1)} on the interval (0, 1). The probability used for
calculating the quantile via \code{\link[=qgeom]{qgeom()}}.}

\item{growth_probability}{\code{numeric(1)} on the interval (0, 1). The
probability that a crop will advance a growth stage when assigned a
random tick.}
}
\description{
the \code{qgeom} function returns an integer value, which can
produce some less than helpful results (such as 10\% of crops will be
mature after 0 seconds). This approximates the geometric distribution
quantiles on a continuous basis.
}
