% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthProbability.R
\name{GrowthProbability}
\alias{GrowthProbability}
\alias{growth_probability}
\title{Calculating Growth Probability}
\usage{
growth_probability(
  ttg,
  method = c("median", "mean"),
  unit = c("tick", "second", "minute", "hour")
)
}
\arguments{
\item{ttg}{\code{numeric(1)}. The stated  time to growth}

\item{method}{\code{character(1)}. The method of calculation. Most Minecraft
resources will report "average time" when using the median. Players should
confirm the calculations make sense before utilizing the results.}

\item{unit}{\code{character(1)}. The unit of measure for the published
time to growth.}
}
\description{
The probability of growth given a random tick is an important
part of calculating metrics related to minecraft farming. The probability
can be found by scouring source code, or sometimes can be found in sources
like the Minecraft Wiki. Frequently, however, the actual growth
probability is not stated, but the "average time to growth" is. These
functions calculate the growth probability for mean and median estimates
of time to growth.
}
