% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_subchunk.R
\name{new_subchunk}
\alias{new_subchunk}
\title{Create Subchunk Object for Simulation}
\usage{
new_subchunk(x = 16, y = 16, z = 16)
}
\arguments{
\item{x, y, z}{\code{integerish(1)} the number of blocks that
make up the x, y, and z dimensions of the subchunk.
Each defaults to 16.}
}
\description{
Creates an object with utilities for
monitoring farm productivity in simulation.
}
