% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerpetualFurnaceIndex.R
\name{PerpetualFurnaceIndex}
\alias{PerpetualFurnaceIndex}
\alias{perpetual_furnace_index}
\title{Perpetual Furnace Index}
\usage{
perpetual_furnace_index(
  total_smelt,
  self_smelt,
  craft_consumption,
  yield,
  growth_probability,
  stages,
  p = 0.5,
  method = c("median", "mean", "quantile")
)
}
\arguments{
\item{total_smelt}{\code{numeric(1)}. The total number of items an item can
smelt. May be a fractional part.}

\item{self_smelt}{\code{numeric(1)}. The number of items of the same type that
must be spent turn the item into fuel.}

\item{craft_consumption}{\code{numeric(1)}. The number of the items used in
crafting a more complex item for fuel.}

\item{growth_probability}{\code{numeric(1)} on the interval (0, 1). The
probability that a crop will advance a growth stage when assigned a
random tick.}

\item{stages}{\code{integerish(1)}. The number of growth stages a crop requires
before becoming mature.}

\item{p}{\code{numeric(1)} on the interval (0, 1). The probability used for
determining the quartile of the probability distribution. Used only
when \code{unit = "minute"}.}

\item{method}{\code{character(1)}. One of \code{c("median", "mean", "quantile")}.
The method of calculation to be based on either the median, mean, or
quantile of the probability distribution.}
}
\description{
The Perpetual Furnace Index (PFI) is the number blocks of
farm that are required to maintain continuous and uninterupted operation
of a single furnace in a self sufficient manner. Furnaces smelt six items
per minute, so this is the number of blocks of farm required to provide
enough fuel to smelt six items per minute. The PFI must also account for
raw materials required to convert itself into fuel
(see \code{\link[=net_smelting_gain]{net_smelting_gain()}}).
}
