% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rtick.R
\name{RandomTicks}
\alias{RandomTicks}
\alias{p_rtick}
\alias{ttrt}
\title{Random Tick Mechanics}
\usage{
p_rtick(
  f,
  fr = 0:n,
  random_tick_speed = getOption("randomTickSpeed", 3),
  B = getOption("subchuckSize", 16^3)
)

ttrt(
  f,
  p = 0.5,
  method = c("median", "mean", "quantile"),
  random_tick_speed = getOption("randomTickSpeed", 3),
  B = getOption("subchunkSize", 16^3)
)
}
\arguments{
\item{f}{\code{integerish(1)}. The number of blocks in the chunk that have
been set aside as farm land.}

\item{fr}{\code{integerish}.  The number of blocks of farmland receiving
a random tick. Greater than 0, but less than \code{n}. If passed as
\code{NULL}, it will not be validated.}

\item{random_tick_speed}{\code{integerish(1)}. The number of blocks in the chunk to receive a
random tick. Defaults to the random tick speed. Must be less than \code{N}.}

\item{B}{\code{integerish(1)}. The number of blocks in a chunk. Defaults to \code{16^3}.}

\item{p}{\code{numeric(1)} on the interval (0, 1). The probability used for
calculating the quantile via \code{\link[=qgeom]{qgeom()}}.}

\item{method}{\code{character(1)}. One of \code{c("median", "mean", "quantile")}.
Permits the user control over whether the median, mean, or a quantile of
Time to Random Tick is returned.}
}
\description{
Functions for Basic Random Tick Mechanics
}
\details{
\code{p_rtick} returns the probability that a random tick is assigned
to a block during a game tick.

\code{ttrt} returns the Time to Random Tick, based on a geometric distribution.
}
