% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickConversion.R
\name{TickConverion}
\alias{TickConverion}
\alias{tick_conversion}
\title{Convert Ticks to Other Time Units}
\usage{
tick_conversion(unit = c("tick", "second", "minute", "hour"))
}
\arguments{
\item{unit}{\code{character(1)}. One of \code{c("tick", "second", "minute", "hour")}.
The unit of time for the metric.}
}
\description{
The fundamental unit of time in Minecraft is the tick.
Each tick lasts 0.05 seconds. Players normally interpret game mechanics
in terms of seconds, minutes, or hours.
}
