% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TickMechanics.R
\name{TickMechanics}
\alias{TickMechanics}
\alias{p_rtick}
\alias{ttrt}
\title{Tick Mechanics}
\usage{
p_rtick(random_tick_speed = getOption("randomTickSpeed", 3))

ttrt(
  random_tick_speed = getOption("randomTickSpeed", 3),
  p = 0.5,
  method = c("median", "mean", "quantile")
)
}
\arguments{
\item{random_tick_speed}{\code{integerish(1)}. The \code{randomTickSpeed} setting for
the world.}

\item{p}{\code{numeric(1)} on the interval (0, 1). The probability used for
calculating the quantile via \code{\link[=qgeom]{qgeom()}}.}

\item{method}{\code{character(1)}. One of \code{c("median", "mean", "quantile")}.
Permits the user control over whether the median, mean, or a quantile of
Time to Random Tick is returned.}
}
\description{
Functions for Basic Tick Mechanics
}
\details{
\code{p_rtick} returns the probability that a random tick is assigned
to a block during a game tick.

\code{ttrt} returns the Time to Random Tick, based on a geometric distribution.
}
