% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeToGrowth.R
\name{TimeToGrowth}
\alias{TimeToGrowth}
\alias{time_to_growth}
\title{Calculate Time To Growth}
\usage{
time_to_growth(
  growth_probability,
  p = 0.5,
  method = c("median", "mean", "quantile"),
  unit = "minute"
)
}
\arguments{
\item{growth_probability}{\code{numeric(1)} on the interval (0, 1). The
probability that a crop will advance a growth stage when assigned a
random tick.}

\item{p}{\code{numeric(1)} on the interval (0, 1). The probability used for
determining the quartile of the probability distribution. Used only
when \code{unit = "minute"}.}

\item{method}{\code{character(1)}. One of \code{c("median", "mean", "quantile")}.
The method of calculation to be based on either the median, mean, or
quantile of the probability distribution.}

\item{unit}{\code{character(1)}. One of \code{c("tick", "second", "minute", "hour")}.
The unit of time for the metric.}
}
\description{
The Time to Growth (TTG) is the duration of time until a crop
advances a growth stage measured in minutes. The probability of growth is
the intersection of the independent events of a random tick being
assigned and growth occurring.
}
