% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_common_farm_args.R
\name{validate_common_farm_args}
\alias{validate_common_farm_args}
\title{Validate Common Farm Arguments}
\usage{
validate_common_farm_args(f, random_tick_speed, B, fr = NULL, coll)
}
\arguments{
\item{f}{\code{integerish(1)}. The number of blocks in the chunk that have
been set aside as farm land.}

\item{random_tick_speed}{\code{integerish(1)}. The number of blocks in the chunk to receive a
random tick. Defaults to the random tick speed. Must be less than \code{N}.}

\item{B}{\code{integerish(1)}. The number of blocks in a chunk. Defaults to \code{16^3}.}

\item{fr}{\code{integerish}.  The number of blocks of farmland receiving
a random tick. Greater than 0, but less than \code{n}. If passed as
\code{NULL}, it will not be validated.}

\item{coll}{An \code{assertCollection} object.}
}
\description{
A utility function to validate the arguments that are commonly
passed with farm calculations, such as farm size, random tick speed,
or subchunk size.
}
