% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToGrowth.R
\name{TimeToGrowth}
\alias{TimeToGrowth}
\alias{ttg}
\title{Time To Growth}
\usage{
ttg(
  p_growth,
  f,
  p = 0.5,
  method = c("median", "mean", "quantile"),
  random_tick_speed = getOption("randomTickSpeed", 3),
  B = getOption("subchunkSize", 16^3)
)
}
\arguments{
\item{f}{\code{integerish(1)}. The number of blocks in the chunk that have
been set aside as farm land.}

\item{p}{\code{numeric(1)} on the interval (0, 1). The probability used for
calculating the quantile via \code{\link[=qgeom]{qgeom()}}.}

\item{method}{\code{character(1)}. One of \code{c("median", "mean", "quantile")}.
Permits the user control over whether the median, mean, or a quantile of
Time to Random Tick is returned.}

\item{random_tick_speed}{\code{integerish(1)}. The number of blocks in the chunk to receive a
random tick. Defaults to the random tick speed. Must be less than \code{N}.}

\item{B}{\code{integerish(1)}. The number of blocks in a chunk. Defaults to \code{16^3}.}
}
\description{
Estimate the time to a growth stage event for a crop in a farm
of a specified size.
}
